timesheet_service/requirements.txt
mysql-connector-python>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0
playwright>=1.40.0
schedule>=1.2.1
